#9 Add routes for different views
