+++ src/styles/header.css
@@ -8,9 +8,10 @@
   justify-content: center;
 }
 
-h1 {
+#headerContainer h1 {
   margin-top: 0;
   margin-bottom: 4px;
+  text-decoration: none;
 }
 
 .button {
@@ -26,6 +27,8 @@ h1 {
   transition-duration: 0.3s;
   cursor: pointer;
   outline: none;
+  padding: 10px;
+  margin-top: 10px;
 }
 
 .button:hover {
